backend/requirements.txt
Django==4.2.7
django-cors-headers==4.3.1
PyJWT==2.8.0
python-decouple==3.8
bcrypt==4.1.2
django-extensions==3.2.3
